implement profit api

+++ server/scraper/src/GemLevelProtScraper/PoeDb/Models.cs
@@ -1,11 +1,5 @@
-using System;
 using System.Collections.Immutable;
-using System.Text.Json;
-using Microsoft.Extensions.Options;
 using MongoDB.Bson.Serialization.Attributes;
-using MongoDB.Driver;
-using MongoDB.Migration;
-using ScrapeAAS;
 
 namespace GemLevelProtScraper.PoeDb;
 
@@ -21,9 +15,9 @@ internal sealed record PoeDbSkillRelatedGem(PoeDbSkillName Name, string Text);
 
 internal sealed record PoeDbSkillDescription(string Text, ImmutableArray<PoeDbSkillRelatedGem> RelatedGems);
 
-internal sealed record PoeDbStatRequirements(double? Intelligence, double? Dexterity, double? Strenght);
+internal sealed record PoeDbStatRequirements(decimal? Intelligence, decimal? Dexterity, decimal? Strenght);
 
-internal sealed record PoeDbSkillLevel(double Level, double RequiresLevel, PoeDbStatRequirements Requirements, double? Experience);
+internal sealed record PoeDbSkillLevel(decimal Level, decimal RequiresLevel, PoeDbStatRequirements Requirements, decimal? Experience);
 
 internal sealed record PoeDbLink(string Label, string Link);
